requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
Deprecated==1.2.13
fonttools==4.38.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.24.0
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
pycparser==2.21
PyGithub==1.57
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
urllib3==1.26.13
wrapt==1.14.1
